backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rashad Philizaire <rashadphilizaire@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
start = "main:start"

[tool.poetry.dependencies]
python = "^3.10"
langchainplus-sdk = "0.0.20"
pydantic-core = "2.1.2"
pydantic = "1.10.9"
celery = "^5.3.5"
fastapi = "^0.104.1"
pinecone-client = "^2.2.4"
scrapy = "^2.11.0"
uvicorn = "^0.24.0.post1"
bs4 = "^0.0.1"
python-dotenv = "^1.0.0"
openai = "^1.3.3"
langchain = "^0.0.338"
tiktoken = "^0.5.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
html2text = "^2020.1.16"
lxml = "^4.9.3"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
